+++ CHANGELOG.md
@@ -68,6 +68,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 * Add marker deposit access check for sends to marker escrow account.  Will break any current address that is sending to the
 marker escrow account if it does not have deposit access.  In order for it to work, deposit access needs to be added.  This can be done using the `MsgAddAccessRequest` tx  [#1525](https://github.com/provenance-io/provenance/issues/1525).
 * `MsgMultiSend` is now limited to a single `Input` [PR 1506](https://github.com/provenance-io/provenance/pull/1506).
+* SDK errors returned from Metadata module endpoints [#978](https://github.com/provenance-io/provenance/issues/978).
 
 ---